last one  commit

+++ app/Http/Controllers/categoryController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\API;
 
+use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Category;
 
@@ -10,34 +11,51 @@ class CategoryController extends Controller
     public function index()
     {
         $categories = Category::all();
-    
-        return response()->json(["data"=>$categories, 200]);
-    
-    } 
+        return response()->json($categories);
+    }
+
+    public function show($category_no)
+    {
+        $category = Category::where('category_no', $category_no)->first();
+        if ($category) {
+            return response()->json($category);
+        } else {
+            return response()->json(['message' => 'Category not found'], 404);
+        }
+    }
 
     public function store(Request $request)
     {
-        $category = Category::create($request->all());
+        $category = new Category;
+        $category->category_name = $request->input('category_name');
+        $category->category_image = $request->input('category_image');
+        $category->save();
 
         return response()->json($category, 201);
     }
 
-    public function show(Category $category)
+    public function update(Request $request, $category_no)
     {
-        return response()->json($category, 200);
+        $category = Category::where('category_no', $category_no)->first();
+        if ($category) {
+            $category->category_name = $request->input('category_name');
+            $category->category_image = $request->input('category_image');
+            $category->save();
+
+            return response()->json($category);
+        } else {
+            return response()->json(['message' => 'Category not found'], 404);
+        }
     }
 
-    public function update(Request $request, Category $category)
+    public function destroy($category_no)
     {
-        $category->update($request->all());
-
-        return response()->json($category, 200);
-    }
-
-    public function destroy(Category $category)
-    {
-        $category->delete();
-
-        return response()->json(null, 204);
+        $category = Category::where('category_no', $category_no)->first();
+        if ($category) {
+            $category->delete();
+            return response()->json(['message' => 'Category deleted']);
+        } else {
+            return response()->json(['message' => 'Category not found'], 404);
+        }
     }
 }
